GEODE-907: Switch all fields from floats to doubles
- This is because when Jackson serializes a float (to JSON) it does so
  through an implicit cast to a double which alters the precision.
  Practically, a number like 23.99F ends up being 23.98999999234D which
  messes with unit tests and displaying in the UI.

+++ geode-pulse/src/main/java/com/vmware/gemfire/tools/pulse/internal/service/ClusterDiskThroughputService.java
@@ -55,8 +55,8 @@ public class ClusterDiskThroughputService implements PulseService {
     // cluster's Throughout Writes trend added to json response object
     // CircularFifoBuffer throughoutWritesTrend =
     // cluster.getThroughoutWritesTrend();
-    Float currentThroughputWrites = cluster.getDiskWritesRate();
-    Float currentThroughputReads = cluster.getDiskReadsRate();
+    double currentThroughputWrites = cluster.getDiskWritesRate();
+    double currentThroughputReads = cluster.getDiskReadsRate();
 
     responseJSON.put("currentThroughputReads", currentThroughputReads);
     responseJSON.put("throughputReads",
